cleaned up files, moved chat-page into its own route so we can further create landing page
