+++ kong-ai/middleware.ts
@@ -2,6 +2,7 @@ import { clerkMiddleware, createRouteMatcher } from '@clerk/nextjs/server'
 
 const isPublicRoute = createRouteMatcher(
   [
+    '/',
     '/sign-in(.*)', 
     '/sign-up(.*)', 
     "/api/assistant/create",
